[refactor] Moved some stuff into utils package
